Add files via upload
This document presents the basic equations in reference to which artificial neural networks are designed and implemented. The scope is restricted to the simpler feed-forward networks, including those having hidden layers. Convolutional and recurrent networks are out of the scope.

Equations are often initially noted using an index-basd notation for the typical element. However, all the equations are eventually cast in the direct matrix form, using a consistent set of notation. Some of the minor aspects of notation were invented to make the presentation as simple and direct as possible.

The presentation here regards a layer as the basic unit. The term “layer” is understood in the same sense in which APIs of modern libraries like TensorFlow-Keras 2.x use it. The presentation here is detailed enough that neural networks with hidden layers could be implemented, starting from the scratch.